requirements.txt
Flask==2.3.0
hvac==1.1.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
